requirements.txt
annotated-types==0.5.0
anyio==3.7.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.1.3
fastapi==0.103.1
google-auth==2.23.0
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
peewee==3.17.0
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==1.26.16
uvicorn==0.23.2
